+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.1.1 (UNRELEASED)
 
 * Add masking support [#1](https://github.com/fpco/amber/issues/1)
+* Add subcommand `generate` [#7](https://github.com/fpco/amber/pull/7)
 
 ## 0.1.0 (2021-08)